🐛:Fix change voice bug.

+++ ZEGOLiveDemo/ZEGOLiveDemo/CorePages/MusicEffects/MusicEffectsVC+BGM.swift
@@ -97,7 +97,7 @@ extension MusicEffectsVC : UICollectionViewDelegate,UICollectionViewDataSource,U
             for model in self.backMusicArr {
                 if index == newIndex {
                     model.isSelected = !model.isSelected
-                    RoomManager.shared.soundService.setBGM(model.selectedType, stop: !model.isSelected)
+                    RoomManager.shared.soundService.setBGM(Int(model.selectedType), stop: !model.isSelected)
                 } else {
                     model.isSelected = false
                 }
@@ -107,7 +107,7 @@ extension MusicEffectsVC : UICollectionViewDelegate,UICollectionViewDataSource,U
             for model in self.voiceChangeArr {
                 if index == newIndex {
                     model.isSelected = true
-                    RoomManager.shared.soundService.setVoiceChangeType(ZegoVoiceChangerPreset(rawValue: UInt(model.selectedType)) ?? .none)
+                    RoomManager.shared.soundService.setVoiceChangeType(ZegoVoiceChangerPreset(rawValue: model.selectedType) ?? .none)
                 } else {
                     model.isSelected = false
                 }
@@ -117,7 +117,7 @@ extension MusicEffectsVC : UICollectionViewDelegate,UICollectionViewDataSource,U
             for model in self.reverberArr {
                 if index == newIndex {
                     model.isSelected = true
-                    RoomManager.shared.soundService.setReverbPreset(ZegoReverbPreset(rawValue: UInt(model.selectedType)) ?? .none)
+                    RoomManager.shared.soundService.setReverbPreset(ZegoReverbPreset(rawValue: model.selectedType) ?? .none)
                 } else {
                     model.isSelected = false
                 }
